fix(sessions): preserve rebased persistence lint budgets
